Include all transactions into the block
Previously we were running an auction to include only the transaction
with the highest fee.

But because the flywheel sends transactions, and people are using both
the flashabke.xyz relay and the baking bad relay, we often have several
transactions in the flashbake endpoint's mempool.

Here we sort them to make sure we only include one transaction per
manager per block, and we pass them all to the baker.

We will reintroduce auction later, but it will be opt-in. The bundle
will have a flag `FIRST_OR_NOTHING` that indicates that it only wants to
be included first, or not at all.